+++ fronts/apps/funeralv2/src/views/system/menu/modules/form.vue
@@ -63,6 +63,9 @@ const schema: VbenFormSchema[] = [
       .max(30, $t('ui.formRules.maxLength', [$t('system.menu.menuName'), 30]))
       .refine(
         async (value: string) => {
+          if (formData.value?.id && value === formData.value.name) {
+            return true;
+          }
           return !(await isMenuNameExists(value, formData.value?.id));
         },
         (value) => ({
@@ -176,6 +179,9 @@ const schema: VbenFormSchema[] = [
       )
       .refine(
         async (value: string) => {
+          if (formData.value?.id && value === formData.value.path) {
+            return true;
+          }
           return !(await isMenuPathExists(value, formData.value?.id));
         },
         (value) => ({
@@ -208,7 +214,7 @@ const schema: VbenFormSchema[] = [
         $t('ui.formRules.startWith', [$t('system.menu.path'), '/']),
       )
       .refine(async (value: string) => {
-        return await isMenuPathExists(value, formData.value?.id);
+        return await isMenuPathExists(value);
       }, $t('system.menu.activePathMustExist'))
       .optional(),
   },