+++ microservices/AuthServer/Services/II18nResourceService.cs
@@ -1,4 +1,5 @@
 using AuthServer.DTOs;
+using static AuthServer.Services.I18nResourceService;
 
 namespace AuthServer.Services;
 
@@ -6,7 +7,7 @@ public interface II18nResourceService
 {
     Task<List<I18nResourceDto>> GetAllResourcesAsync();
     Task<List<I18nResourceDto>> GetResourcesByLocaleAsync(string locale);
-    Task<List<I18nResourceDto>> GetPagedResourcesAsync(SearchI18nParams searchParams);
+    Task<PagedResourceDto<I18nResourceDto>> GetPagedResourcesAsync(SearchI18nParams searchParams);
     Task<I18nResourceDto> CreateResourceAsync(CreateI18nResourceDto request);
     Task<bool> UpdateResourceAsync(int id, CreateI18nResourceDto request);
     Task<bool> DeleteResourceAsync(int id);